Wallet: move sync code to module
